Make IOS 11+ compatible
I made the IOS 7-10 settings default for all IOS devices but left the styles customisable for specific IOS versions. (as in the case of IOS6).

This should stop the code breaking when new IOS versions are released, but still offer the option to make custom changes if the new IOS's UI changes. 

Note: I have NOT tested the on all IOS visions devices.